development: advice to launch XCode after installing it (bgo#750551)
It can happen that after installing XCode, not all of the
development tools that are needed were completely installed.
After trying to launch it, it can happen that the program will
prompt you to install more tools, which can fix other problems
down the line, like the error mentioned in bug 750551.