development: move a small paragraph to prevent disordered list
The numbered bullet points in this markdown list somehow were
getting corrupted when translated to HTML code, because the last
item ("5. Build Banshee") appeared as "1. Build Banshee".

This seems to happen, somehow, if a paragraph is written below
one of the bullet points (it makes the <ol> element close, and
open a new one for a new bullet point). So one way to fix this
is place the contents of this paragraph as part of the bullet
item text that is numbered.